% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_objectives.R
\name{add_objective_max_profit}
\alias{add_objective_max_profit}
\title{Add objective: maximize profit}
\usage{
add_objective_max_profit(
  x,
  profit_col = "profit",
  actions = NULL,
  alias = NULL
)
}
\arguments{
\item{x}{A \code{Data} object.}

\item{profit_col}{Character. Profit column in \code{x$data$dist_profit}.}

\item{actions}{Optional subset of actions to include. Values may match \code{actions$id}
and, if present, \code{actions$action_set}.}

\item{alias}{Optional identifier for multi-objective workflows.}
}
\value{
Updated \code{Data} object.
}
\description{
Specify an objective that maximizes total profit from selected \verb{(pu, action)} pairs.
The objective can optionally be restricted to a subset of actions.
}
