% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_objectives.R
\name{add_objective_min_fragmentation}
\alias{add_objective_min_fragmentation}
\title{Add objective: minimize fragmentation}
\usage{
add_objective_min_fragmentation(
  x,
  relation_name = "boundary",
  weight_multiplier = 1,
  alias = NULL
)
}
\arguments{
\item{x}{A \code{Data} object.}

\item{relation_name}{Character. Name of the spatial relation.}

\item{weight_multiplier}{Numeric >= 0. Multiplier applied to relation weights.}

\item{alias}{Optional identifier for multi-objective workflows.}
}
\value{
Updated \code{Data} object.
}
\description{
Specify an objective that minimizes planning-unit fragmentation over a spatial relation.
}
