% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_group_equity.R
\name{add_objective_min_group_mean}
\alias{add_objective_min_group_mean}
\title{Minimize the mean relative contribution across groups}
\usage{
add_objective_min_group_mean(x, alias = NULL)
}
\arguments{
\item{x}{A \link{Problem} configured with \code{\link[=add_group_equity]{add_group_equity()}}.}

\item{alias}{Optional objective alias for multi-objective methods.}
}
\value{
A modified \link{Problem} object.
}
\description{
Minimize the mean relative contribution across groups
}
