% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_sim.R
\name{load_sim_features_raster}
\alias{load_sim_features_raster}
\title{Example feature raster}
\usage{
load_sim_features_raster()
}
\value{
A \code{terra::SpatRaster}.
}
\description{
Load the example feature raster shipped with the package.
}
