% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-sim.R
\docType{data}
\name{sim_pu}
\alias{sim_pu}
\title{Simulated planning units}
\format{
An object of class \code{sf}.
}
\usage{
data(sim_pu)
}
\description{
Example planning units as an \code{sf} object for package examples and tests.
}
\keyword{datasets}
